Add test case for duplicated fragments
