+++ packages/presets/client/tests/client-preset.spec.ts
@@ -4,6 +4,7 @@ import { executeCodegen } from '@graphql-codegen/cli';
 import { mergeOutputs } from '@graphql-codegen/plugin-helpers';
 import { validateTs } from '@graphql-codegen/testing';
 import { preset } from '../src/index.js';
+import { print } from 'graphql';
 
 describe('client-preset', () => {
   it('can generate simple examples uppercase names', async () => {
@@ -999,7 +1000,13 @@ export * from "./gql.js";`);
     expect(graphqlFile).toBeDefined();
   });
 
-  it.only('should dedupe fragments - #8670', async () => {
+  it('should dedupe fragments - #8670', async () => {
+    const dir = path.join(__dirname, 'tmp/duplicate-fragments');
+    const cleanUp = async () => {
+      await fs.promises.rm(dir, { recursive: true, force: true });
+    };
+
+    const docPath = path.join(__dirname, 'fixtures/reused-fragment.ts');
     const result = await executeCodegen({
       schema: [
         /* GraphQL */ `
@@ -1021,29 +1028,7 @@ export * from "./gql.js";`);
           }
         `,
       ],
-      documents: [
-        /* GraphQL */ `
-          fragment SharedComponentFragment on User {
-            id
-            username
-          }
-
-          fragment EventHeaderComponentFragment on Event {
-            owner {
-              ...SharedComponentFragment
-            }
-          }
-
-          query EventQuery($eventId: ID!) {
-            event(id: $eventId) {
-              ...EventHeaderComponentFragment
-              attendees {
-                ...SharedComponentFragment
-              }
-            }
-          }
-        `,
-      ],
+      documents: [docPath],
       generates: {
         'out1/': {
           preset,
@@ -1051,7 +1036,26 @@ export * from "./gql.js";`);
         },
       },
     });
-    const output = mergeOutputs([...result]);
+
+    // TODO: Consider using in-memory file system for tests like this.
+    try {
+      await cleanUp();
+    } catch {}
+    await fs.promises.mkdir(path.join(dir, 'out1'), { recursive: true });
+    for (const file of result) {
+      await fs.promises.writeFile(path.join(dir, file.filename), file.content, 'utf-8');
+    }
+
+    const { default: jiti } = await import('jiti');
+    const loader = jiti('', {});
+
+    const { EventQueryDocument } = loader(path.join(dir, 'out1/graphql.ts'));
+
+    const printed = print(EventQueryDocument);
+
+    expect(printed.match(/fragment SharedComponentFragment on User/g)?.length).toBe(1);
+
+    await cleanUp();
   });
 
   describe('when no operations are found', () => {
